WIP: split firmware in two
This has quite a few rough edges
