+++ launcher/assets.yml
