+++ launcher/sprites.png
